Move `HasFields` trait to separate module
